# app/readium-backend/src/main/java/com/br/klaus/readium/translation/dto/AutoTranslationRequestDTO.java
package com.br.klaus.readium.translation.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record AutoTranslationRequestDTO(
        @NotBlank
        String text,
        @NotBlank
        @Size(max = 10)
        String targetLanguage
) {
}